requirements.txt
asgiref==3.2.5
Django==2.2.11
django-cors-headers==3.2.1
django-environ==0.4.5
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.1
